requirements.txt
pytest-playwright
pytest-html